backend/requirements.txt
fastapi
uvicorn
pyserial
requests
flask
flask-socketio
flask-cors